+++ src/main/java/com/robertx22/age_of_exile/database/data/currency/ClearSocketsItem.java
@@ -38,7 +38,6 @@ public class ClearSocketsItem extends CurrencyItem implements ICurrencyItemEffec
     public ItemStack ModifyItem(ItemStack stack, ItemStack Currency) {
         GearItemData gear = Gear.Load(stack);
         gear.sockets.sockets.clear();
-        gear.sockets.word = "";
         Gear.Save(stack, gear);
         return stack;
     }
@@ -47,8 +46,6 @@ public class ClearSocketsItem extends CurrencyItem implements ICurrencyItemEffec
     public List<BaseLocRequirement> requirements() {
         return Arrays.asList(GearReq.INSTANCE);
     }
-
-    
 
     @Override
     public float getInstability() {