+++ src/main/java/com/robertx22/age_of_exile/uncommon/datasaving/Gear.java
@@ -56,7 +56,6 @@ public class Gear {
             stack.setTag(new NbtCompound());
         }
         if (gear != null) {
-            gear.sockets.setSocketsCount(gear);
 
             LoadSave.Save(gear, stack.getTag(), LOC);
         }